Restructure test files into separate directory, update one of the tests
