+++ tests/testfiles/both-arg-types.spec.js
@@ -1,4 +1,4 @@
-const runnableExports = require('./../src/index.js')
+const runnableExports = require('../../src/index.js')
 
 module.exports.test = (objectArg, ...arrayArgs) => {
 	console.log(objectArg, arrayArgs)
