+++ tests/testfiles/defaultexport.spec.js
@@ -1,4 +1,4 @@
-const runnableExports = require('./../src/index.js')
+const runnableExports = require('../../src/index.js')
 
 module.exports = (...testArgs) => {
 	console.log(testArgs)
